+++ mopack/builders/cmake.py
@@ -1,11 +1,12 @@
 import os.path
 
-from . import Builder, BuilderOptions
+from . import ConfiguringBuilder, BuilderOptions
+from .ninja import NinjaBuilder
 from .. import types
 from ..environment import get_cmd
 from ..freezedried import GenericFreezeDried
 from ..log import LogFile
-from ..path import pushd
+from ..path import Path, pushd
 from ..shell import ShellArguments
 
 # XXX: Handle exec-prefix, which CMake doesn't work with directly.
@@ -13,9 +14,9 @@ _known_install_types = ('prefix', 'bindir', 'libdir', 'includedir')
 
 
 @GenericFreezeDried.fields(rehydrate={'extra_args': ShellArguments})
-class CMakeBuilder(Builder):
+class CMakeBuilder(ConfiguringBuilder):
     type = 'cmake'
-    _version = 2
+    _version = 3
 
     class Options(BuilderOptions):
         type = 'cmake'
@@ -40,10 +41,16 @@ class CMakeBuilder(Builder):
         # v2 removes the name field.
         if version < 2:  # pragma: no branch
             del config['name']
+
+        # v3 adds `directory`.
+        if version < 3:  # pragma: no branch
+            config['directory'] = Path('', Path.Base.srcdir).dehydrate()
+
         return config
 
     def __init__(self, pkg, *, extra_args=None, _symbols, **kwargs):
-        super().__init__(pkg, **kwargs)
+        super().__init__(pkg, _symbols=_symbols, _child_builder=NinjaBuilder,
+                         **kwargs)
 
         _symbols = _symbols.augment_path_bases(*self.path_bases())
         T = types.TypeCheck(locals(), _symbols)
@@ -65,24 +72,14 @@ class CMakeBuilder(Builder):
 
         env = self._common_options.env
         cmake = get_cmd(env, 'CMAKE', 'cmake')
-        ninja = get_cmd(env, 'NINJA', 'ninja')
         with LogFile.open(metadata.pkgdir, self.name) as logfile:
             with pushd(path_values['builddir'], makedirs=True, exist_ok=True):
                 logfile.check_call(
-                    cmake + [path_values['srcdir'], '-G', 'Ninja'] +
+                    cmake + [self.directory.string(**path_values),
+                             '-G', 'Ninja'] +
                     self._toolchain_args(self._this_options.toolchain) +
                     self._install_args(self._common_options.deploy_dirs) +
                     self.extra_args.fill(**path_values),
                     env=env
                 )
-                logfile.check_call(ninja, env=env)
-
-    def deploy(self, metadata, pkg):
-        path_values = pkg.path_values(metadata)
-
-        env = self._common_options.env
-        ninja = get_cmd(env, 'NINJA', 'ninja')
-        with LogFile.open(metadata.pkgdir, self.name,
-                          kind='deploy') as logfile:
-            with pushd(path_values['builddir']):
-                logfile.check_call(ninja + ['install'], env=env)
+        super().build(metadata, pkg)