+++ test/unit/origins/sdist/test_tarball.py
@@ -338,6 +338,7 @@ class TestTarball(SDistTestCase):
 
         with mock_open_log() as mopen, \
              mock.patch('mopack.builders.bfg9000.pushd'), \
+             mock.patch('mopack.builders.ninja.pushd'), \
              mock.patch('subprocess.run') as mrun:
             with assert_logging([('resolve', 'foo')]):
                 pkg.resolve(self.metadata)
@@ -353,6 +354,7 @@ class TestTarball(SDistTestCase):
 
         with mock_open_log() as mopen, \
              mock.patch('mopack.builders.bfg9000.pushd'), \
+             mock.patch('mopack.builders.ninja.pushd'), \
              mock.patch('subprocess.run'):
             with assert_logging([('deploy', 'foo')]):
                 pkg.deploy(self.metadata)