+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## v0.2.0 (in progress)
 
+### New features
+- `build` field for source distributions can now take a list of builders
+- New builder `ninja`
+
 ### Breaking changes
 - `usage`/`usages` are now `linkage`/`linkages`