+++ doc/about/changes.md
@@ -4,6 +4,10 @@
 in progress
 {: .subtitle}
 
+### New features
+- `build` field for source distributions can now take a list of builders
+- New builder `ninja`
+
 ### Breaking changes
 - `usage`/`usages` are now `linkage`/`linkages`